refactor: event handler & TransactionalEventListener 도입
