+++ backend/src/main/java/com/deliveranything/domain/delivery/handler/redis/OrderDeliveryStatusRedisSubscriber.java
@@ -1,4 +1,4 @@
-package com.deliveranything.domain.delivery.handler;
+package com.deliveranything.domain.delivery.handler.redis;
 
 import com.deliveranything.domain.delivery.entity.Delivery;
 import com.deliveranything.domain.delivery.event.dto.OrderStatusUpdateEvent;
