+++ backend/src/main/java/com/deliveranything/domain/delivery/service/DeliveryService.java
@@ -7,8 +7,8 @@ import com.deliveranything.domain.delivery.entity.Delivery;
 import com.deliveranything.domain.delivery.enums.DeliveryStatus;
 import com.deliveranything.domain.delivery.event.dto.DeliveryStatusEvent;
 import com.deliveranything.domain.delivery.event.dto.OrderStatusUpdateEvent;
-import com.deliveranything.domain.delivery.event.event.redis.DeliveryStatusRedisPublisher;
-import com.deliveranything.domain.delivery.event.event.redis.OrderDeliveryStatusRedisPublisher;
+import com.deliveranything.domain.delivery.handler.DeliveryStatusEventHandler;
+import com.deliveranything.domain.delivery.handler.OrderDeliveryStatusEventHandler;
 import com.deliveranything.domain.delivery.repository.DeliveryRepository;
 import com.deliveranything.domain.order.entity.Order;
 import com.deliveranything.domain.user.entity.profile.RiderProfile;
@@ -27,8 +27,8 @@ public class DeliveryService {
 
   private final RiderProfileService riderProfileService;
   private final DeliveryRepository deliveryRepository;
-  private final DeliveryStatusRedisPublisher deliveryStatusRedisPublisher;
-  private final OrderDeliveryStatusRedisPublisher orderDeliveryStatusRedisPublisher;
+  private final DeliveryStatusEventHandler deliveryStatusEventHandler;
+  private final OrderDeliveryStatusEventHandler orderDeliveryStatusEventHandler;
   private final RedisTemplate<String, Object> redisTemplate;
 
   public void updateRiderStatus(Long riderId, RiderToggleStatusRequestDto riderStatusRequestDto) {
@@ -59,7 +59,7 @@ public class DeliveryService {
         .nextStatus(next)
         .build();
 
-    deliveryStatusRedisPublisher.publish(event);
+    deliveryStatusEventHandler.handleDeliveryStatus(event);
   }
 
   // 라이더 배달 수락/거절 처리 및 상태 이벤트 발행
@@ -70,7 +70,7 @@ public class DeliveryService {
     // 이벤트만 발행 - 실제 상태 변경은 구독자에서 처리
     OrderStatusUpdateEvent event = new OrderStatusUpdateEvent(
         decisionRequestDto.orderId(), currentActiveProfileId, status);
-    orderDeliveryStatusRedisPublisher.publish(event);
+    orderDeliveryStatusEventHandler.handleOrderDeliveryStatus(event);
   }
 
   // Delivery 생성
